Discontinuation of support for Es 7.x and earlier than 8.10
