--- src/main/ext/es-7.15-lt/xcontent-aliases.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) 2024 Works Applications Co., Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@file:Suppress("PackageDirectoryMismatch")
-
-package com.worksap.nlp.lucene.aliases
-
-typealias ToXContent = org.elasticsearch.common.xcontent.ToXContent
-
-typealias ToXContentParams = org.elasticsearch.common.xcontent.ToXContent.Params
-
-typealias XContentBuilder = org.elasticsearch.common.xcontent.XContentBuilder
