+++ src/main/ext/es-8.10-ge/factory-adapters.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2024 Works Applications Co., Ltd.
+ * Copyright (c) 2022-2025 Works Applications Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -34,7 +34,7 @@ abstract class AbstractTokenFilterFactory(
     environment: Environment?,
     name: String?,
     settings: Settings?
-) : org.elasticsearch.index.analysis.AbstractTokenFilterFactory(indexSettings, name, settings)
+) : org.elasticsearch.index.analysis.AbstractTokenFilterFactory(name, settings)
 
 abstract class AbstractIndexAnalyzerProvider<T : Analyzer>(
     indexSettings: IndexSettings?,
