+++ src/main/java/com/worksap/nlp/lucene/aliases/lucene-aliases.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Works Applications Co., Ltd.
+ * Copyright (c) 2023-2025 Works Applications Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-package com.worksap.nlp.lucene.sudachi.ja
+@file:Suppress("PackageDirectoryMismatch")
 
-import org.apache.lucene.document.Document
-import org.apache.lucene.search.IndexSearcher
+package com.worksap.nlp.lucene.aliases
 
-fun TestSudachiAnalyzer.getDocumentFromSearcher(searcher: IndexSearcher, docId: Int): Document {
-  return searcher.getIndexReader().storedFields().document(docId)
-}
+typealias TokenFilterFactory = org.apache.lucene.analysis.TokenFilterFactory
+
+typealias ResourceLoaderArgument = org.apache.lucene.util.ResourceLoader
+
+typealias ResourceLoaderAware = org.apache.lucene.util.ResourceLoaderAware
