--- src/test/ext/es-8.00-lt/engine-aliases-test.kt
@@ -1,35 +1,0 @@
-/*
- * Copyright (c) 2023 Works Applications Co., Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@file:Suppress("PackageDirectoryMismatch")
-
-package com.worksap.nlp.search.aliases
-
-typealias Version = org.elasticsearch.Version
-
-typealias LogConfigurator = org.elasticsearch.common.logging.LogConfigurator
-
-typealias TestEnvironment = org.elasticsearch.env.TestEnvironment
-
-typealias Index = org.elasticsearch.index.Index
-
-typealias AnalysisModule = org.elasticsearch.indices.analysis.AnalysisModule
-
-typealias IndexSettingsModule = org.elasticsearch.test.IndexSettingsModule
-
-typealias NamedAnalyzer = org.elasticsearch.index.analysis.NamedAnalyzer
-
-typealias IndexAnalyzers = org.elasticsearch.index.analysis.IndexAnalyzers
