+++ src/test/ext/es-8.10-ge/metadata-aliases.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Works Applications Co., Ltd.
+ * Copyright (c) 2022-2025 Works Applications Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
+@file:Suppress("UNUSED_PARAMETER", "PackageDirectoryMismatch")
+
 package com.worksap.nlp.elasticsearch.sudachi.aliases
 
-import org.elasticsearch.cluster.metadata.IndexMetaData
+import org.elasticsearch.cluster.metadata.IndexMetadata
 
 object MetadataConstants {
-  const val SETTING_VERSION_CREATED = IndexMetaData.SETTING_VERSION_CREATED
+  const val SETTING_VERSION_CREATED = IndexMetadata.SETTING_VERSION_CREATED
 }
