--- src/test/ext/es-8.10-lt/SearchEngineEnv.kt
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2023-2024 Works Applications Co., Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.worksap.nlp.elasticsearch.sudachi.index
-
-import com.worksap.nlp.search.aliases.*
-
-fun SearchEngineEnv.indexAnalyzers(settings: Settings): IndexAnalyzers {
-  val indexSettings = IndexSettingsModule.newIndexSettings(Index("test", "_na_"), settings)
-  return analysisRegistry.build(indexSettings)
-}
