+++ subplugin/src/main/ext/es-7.04-ge/com/worksap/nlp/elasticsearch/sudachi/plugin/SudachiTestSubplugin.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Works Applications Co., Ltd.
+ * Copyright (c) 2023-2025 Works Applications Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
